Refactor storage client usage in publicLink and WasScreen to eliminate require cycle
