+++ app/lib/publicLink.ts
@@ -7,17 +7,9 @@ import { credentialIdFor, educationalOperationalCredentialFrom } from './decode'
 import * as verifierInstance from './verifierInstance';
 import { StoreCredentialResult } from './verifierInstance';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-// Dynamic import to avoid cycle
-let getStorageClient: any = null;
-const getStorageClientLazy = async () => {
-  if (!getStorageClient) {
-    const wasScreen = await import('../screens/WAS/WasScreen');
-    getStorageClient = wasScreen.getStorageClient;
-  }
-  return getStorageClient;
-};
 import { v4 as uuidv4 } from 'uuid';
 import { WAS, VERIFIER_INSTANCE_URL } from '../../app.config';
+import { getStorageClient } from './storageClient';
 import { getRootSigner } from './getRootSigner';
 import { ISigner } from '@digitalcredentials/ssi';
 
@@ -79,8 +71,7 @@ async function createWasPublicLinkIfAvailable(
 
     const spaceId = `urn:uuid:${storedSpaceUUID}` as `urn:uuid:${string}`;
 
-    const getStorageClientFn = await getStorageClientLazy();
-    const storage = getStorageClientFn();
+    const storage = getStorageClient();
 
     const space = storage.space({
       signer,
@@ -170,8 +161,7 @@ export async function unshareCredential(rawCredentialRecord: CredentialRecordRaw
         const storedSpaceUUID = await AsyncStorage.getItem(WAS.KEYS.SPACE_ID);
         if (storedSpaceUUID) {
           const spaceId = `urn:uuid:${storedSpaceUUID}` as `urn:uuid:${string}`;
-          const getStorageClientFn = await getStorageClientLazy();
-          const storage = getStorageClientFn();
+          const storage = getStorageClient();
           const space = storage.space({
             signer,
             id: spaceId
